# Branchless computation of the integer absolute value