… boot/ocamlc
… boot/ocamldep
… boot/ocamllex
… bytecomp/printlambda.ml