This patch makes `abs` a built-in operation with a more efficient implementation.

A bit of background: one program of mine involves computing [Manhattan distances](https://en.wikipedia.org/wiki/Taxicab_geometry) inside tight loops, and `abs` function is used at the heart of it. While studying the performance of this program, I found out that `Pervasives.abs` has a naïve implementation based on a conditional branch, and that it's a source of frequent branch mis-predictions (according to Valgrind), which should result in a performance hit on modern OoO architectures. Indeed -- replacing `abs` with an optimized builtin (based on a branchless bitmask trick) made my program run 1.5 times as fast on a recent Intel CPU.

A simple benchmark to demonstrate the performance differences:
``` ocaml
let () =
  let sum = ref 0 in
  for i = -100_000_000 to 100_000_000 do
    for j = 52 to 62 do
      sum := !sum + abs (i lsl j);
    done
  done;
  Printf.printf "%d\n" !sum
```

The benchmark is designed to trigger cache misses, without trying too hard.

Results on my PC (i5 Broadwell), taking the minimum of 10 runs:

-- | Before | After |
--|--|--|
ocamlc | 69.524s | 37.688s | 
ocamlopt | 9.404s | 4.280s |

So, the benchmark now runs about twice as fast, which is significant.

I haven't used the more complicated algorithm in the byte-code interpreter, because the benefits of it seem to get lost in the noise of interpretive overhead. The program still runs twice as fast, because there are less opcodes to process, and because there is no function call now.

`ocamlopt` now produces nice code for `abs` on AMD64:
```
sarq $63, %rdx
addq %rdx, %rsi
addq %rdx, %rsi
addq %rdx, %rsi
xorq %rsi, %rdx
```

To be honest, I have no clue how good will it work on non-x86 platforms. Perhaps I should have added an `Iabs` constructor to `Mach` and pushed the implementation further into `amd64` and `i386` backends?

One more thing -- the branchless `abs` computation algorithm that I used here [is patented](http://patents.justia.com/patent/6073150) on behalf of Sun Microsystems, and is probably now a part of Oracle's patent portfolio. Please use the button in the top right corner of this PR form to cast a vote:

- Press :laughing: if you're not afraid of Oracle and their nasty lawyers.
- Press :confused: if you think that this legal issue is to be taken seriously.
- Press :heart: if you like me.

Thank you!