@xavierleroy 2017-03-19
I'm with Linus Torvalds here: never do a patent search when you're working on an open source project.  Damages are much higher when you knowingly infringe on a patent than when you unknowingly infringe.  By documenting the infringement publicly, and making it easy for lawyers to locate the infrigement by a simple Web search, you've made this PR impossible to accept.  Closing!