@cakeplus 2017-03-19
Understood! Sorry.

The patent is totally bogus though, and variations of this very algorithm are used in several language implementations (e.g. Visual Studio).