@xavierleroy 2017-03-19
Now a few more remarks.  First, Hacker's Delight lists 3 branchless sequences for absolute value, and there are probably more, and with high probability not all of them are patented.  (But, again, don't do patent searches yourself.)

Second, adding a new primitive and pushing it to all code emitters is a lot of work: at 20+ modified files you're still not done.  Even more so given that you don't use processor-specific features (e.g. conditional moves, predicated instructions). Try to implement your magic formula in OCaml itself, inside pervasives.ml.  ocamlopt will do a half-assed job at it, with a bit of tagging remaining, but it will still be branchless.  Also, bytecode will suffer...

Third, the benchmark is useful but for completeness you should also benchmark the case where branch prediction is perfect, i.e. the sign of the argument is always the same.