@xavierleroy 2017-03-19
The bogusness of a patent doesn't matter unless you're big enough and rich enough to actually fight the patent in court.  For small fry like us the only viable strategy is first to stay under the radar, then to claim "we didn't know" in good faith.