@cakeplus 2017-03-19
> Try to implement your magic formula in OCaml itself, inside pervasives.ml. ocamlopt will do a half-assed job at it, with a bit of tagging remaining, but it will still be branchless. Also, bytecode will suffer...

That was the first thing I tried and it truly wasn't too bad (6 instructions instead of 5). Bytecode suffering was the reason that motivated me to add a compiler primitive (as `abs` seemed fundamental enough to warrant such move). I'll think more about all this. Thanks for the remarks!

> The bogusness of a patent doesn't matter unless you're big enough and rich enough to actually fight the patent in court. For small fry like us the only viable strategy is first to stay under the radar, then to claim "we didn't know" in good faith.

Makes total sense. I apologize again.