@xavierleroy 2017-03-19
Re: bytecode suffering, you could have your cake (pun intended) and eat it too as follows:
```
let abs =
  if Sys.backend_type = Sys.Native then
      (fun x -> let y = x asr (Sys.word_size - 1) in (x lxor y) - y)
  else
      (fun x -> if x >= 0 then x else -x)
```
ocamlopt is good enough to specialize on `Sys.backend_type` and `Sys.word_size`, giving the same code as
```
let abs x = let y = x asr 63 in (x lxor y) - y)
```
(on a 64-bit platform, say).  At the same time, in bytecode, you just pay the test at start-up time, and applications are as fast as with the current definition of `abs`.