@cakeplus 2017-03-19
> if Sys.backend_type = Sys.Native then

Oh yes! This is a great suggestion. I was thinking of conditional compilation, and this is exactly it, in disguise.

For the record, I've done an alternative benchmark with a perfectly predictable branch:
```ocaml
let () =
  let sum = ref 0 in
  for i = 0 to 1_000_000_000 do
    sum := !sum + abs i;
  done;
  Printf.printf "%d\n" !sum
```
-- | Before | After |
--|--|--|
ocamlc | 23.128s |  13.408s |
ocamlopt | 0.904s | 1.148s |

Surely enough, there is a slight regression with `ocamlopt`.

The big question is: what are the typical use-cases of `abs`? I've just grepped through a big pile of sources in various languages that happened to reside on my hard drive and studied the uses of `abs`. Surprisingly, I only found two kinds of uses:
1) Distance computation (`abs (a - b)`). Very often (dozens of projects).
2) Simplification of rationals (`let simplify (a, b) = let g = abs (gcd a b) in (a / g, b / g)`). Quite rare (three programs).
  
Based on these questionable data points, I make an educated guess that `abs` is typically used in situations where the argument can be either positive or negative with comparable probability, and a branchless implementation of `abs` can bring substantial benefits. Those who read this: what are your typical uses of `abs`? I'm genuinely curious.