@cakeplus 2017-03-19
More on the damned patent (that will haunt me till the end of my days, no less):

I've studied it very carefully and... turns out that the described algorithm is completely different from the one that I've implemented here!

The algorithm from the paper ("enterprise abs") is:
1) Arithmetic shift;
2) Xor;
3) Subtraction.

And my algorithm ("tagged bit abs") is:
1) Arithmetic shift;
2) Addition;
3) Addition;
4) Addition;
5) Xor.

As should be obvious to anyone reading this, it is the ideological opposite of "enterprise abs". Not only the core idea is turned upside down, it also contains a novel groundbreaking solution that makes it applicable to bit-tagged integers!

If I were an enterprise jerk, I'd surely file a patent.