@cakeplus 2017-03-19
Regarding invasiveness of the patch: one thing that hadn't occurred to me earlier is simply adding a new C primitive ("caml_absint"). It would also allow to make use of `abs` from the C standard library.