@xavierleroy 2017-03-20
Calling a C function (even of the "noalloc" kind) is more expensive than evaluating the original OCaml definition of "abs".  Remember, C primitives are not inlined, while OCaml definitions are.