@chambart 2017-03-23
By the way, if you want to play with other solutions, there is the predicated version:
```assembly
        movq    $2, %rbx
        subq    %rax, %rbx
        cmovnl  %rbx, %rax
```
It seems to be better than both the if and bit-fiddling versions on both well and badly predicted cases.
It could somewhat be done as a general optimization on if. Yet it would be tricky to be able to use
`subq` for both the computation and the conditional in a general way.