@cakeplus 2017-03-23
Good suggestion, @chambart. This is what I came up with on ARM, using the same core idea that plays well with bit-tagging:
```asm
CMP r0, #0
RSBLT r0, r0, #2 ; if r0 is less than zero, set r0 to 2 - r0
```

I've planned an alternative PR with a thrice larger scope named "Branchless computation of abs, min, and max" that will compare:
1) The hardcore approach of adding an "abstract cmov" operation to `Mach` and extending all code emitters. I'm currently having high-quality entertainment with this.
2) The simple approach of putting the branchless formulas right into `Pervasives`, as suggested by @xavierleroy. With `ocamlopt`, on x86, it's about as good (performance-wise) as the solution proposed in this PR.

I haven't decided yet for which approach I will argue (there are more factors to consider beside performance), but you'll get the whole picture.